pyproject.toml
[project]
name = "skillprobe"
version = "0.2.0"
description = "Automated end-to-end skill testing for LLM coding tools"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
skillprobe = "skillprobe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/skillprobe"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
